+++ src/plugins/weechat-plugin.h
@@ -74,7 +74,7 @@ struct t_weelist_item;
  * please change the date with current one; for a second change at same
  * date, increment the 01, otherwise please keep 01.
  */
-#define WEECHAT_PLUGIN_API_VERSION "20240402-01"
+#define WEECHAT_PLUGIN_API_VERSION "20240610-01"
 
 /* macros for defining plugin infos */
 #define WEECHAT_PLUGIN_NAME(__name)                                     \
@@ -138,6 +138,7 @@ struct t_weelist_item;
 #define WEECHAT_HASHTABLE_POINTER                   "pointer"
 #define WEECHAT_HASHTABLE_BUFFER                    "buffer"
 #define WEECHAT_HASHTABLE_TIME                      "time"
+#define WEECHAT_HASHTABLE_LONGLONG                  "longlong"
 
 /* types for hdata */
 enum t_weechat_hdata